Fix: /component/entry should be supported still
Oceanify@5 introduces a lot of breaking changes. One of them is that the
components, both css and js, are now referenced with full path in the
format of `/${name}/${version}/${entry}`. This change makes versioning
of components possible.

The change itself is not backward compatible, which makes legacy
code hard to upgrade. Hence with this commit, we brought the good old
`/${entry}` reference of component back.

Also to be warned, the components still will be compiled into
`${dest}/${name}/${version}/` directory. This commit only brings the dev
behaviour back.