fix: porter.pack() should ignore incomplete modules
for example, when requesting /entry.css and /entry.js simutaneously, and porter starts packing bundles for entry.css, entry.js might be added to porter.entries already but the module isn't fully parsed yet. It should be ignored, and left to the second packing iteration.

Otherwise the packed entry.js is incomplete